moving code around and some renaming
