+++ ec/ecresource/deploymentresource/delete.go
@@ -21,12 +21,9 @@ import (
 	"context"
 	"errors"
 
-	"github.com/elastic/cloud-sdk-go/pkg/api"
 	"github.com/elastic/cloud-sdk-go/pkg/api/deploymentapi"
-	"github.com/elastic/cloud-sdk-go/pkg/api/deploymentapi/trafficfilterapi"
 	"github.com/elastic/cloud-sdk-go/pkg/client/deployments"
 	deploymentv2 "github.com/elastic/terraform-provider-ec/ec/ecresource/deploymentresource/deployment/v2"
-	"github.com/elastic/terraform-provider-ec/ec/internal/util"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 )
 
@@ -72,32 +69,4 @@ func (r *Resource) Delete(ctx context.Context, req resource.DeleteRequest, resp 
 func alreadyDestroyed(err error) bool {
 	var destroyed *deployments.ShutdownDeploymentNotFound
 	return errors.As(err, &destroyed)
-}
-
-func removeRule(ruleID, deploymentID string, client *api.API) error {
-	res, err := trafficfilterapi.Get(trafficfilterapi.GetParams{
-		API: client, ID: ruleID, IncludeAssociations: true,
-	})
-
-	// If the rule is gone (403 or 404), return nil.
-	if err != nil {
-		if util.TrafficFilterNotFound(err) {
-			return nil
-		}
-		return err
-	}
-
-	// If the rule is found, then delete the association.
-	for _, assoc := range res.Associations {
-		if deploymentID == *assoc.ID {
-			return trafficfilterapi.DeleteAssociation(trafficfilterapi.DeleteAssociationParams{
-				API:        client,
-				ID:         ruleID,
-				EntityID:   *assoc.ID,
-				EntityType: *assoc.EntityType,
-			})
-		}
-	}
-
-	return nil
 }
